## Summary
- **Security**: FCM token type validation (prevents Firestore object
injection), log entry type/length validation
- **Bug fix**: Duplicate `const newCoins` in economy.js causing
SyntaxError
- **Test hardening**: Fixed 7 pre-existing test failures (conversations
+ device-info mock gaps), added 19 new regression tests, removed
duplicate test
- **Code quality**: InputStream leak fixes (7 Kotlin files), XML comment
syntax (19 locales), JSDoc headers, semantic HTML, Cache-Control
headers, responsive design fixes

## Key changes
| Category | Count |
|----------|-------|
| Security fixes | 2 (FCM token injection, log field injection) |
| Bug fixes | 1 (economy.js syntax error) |
| Test failures fixed | 7 (pre-existing mock issues) |
| New tests added | 19 (notifications + logs routes) |
| Resource leak fixes | 7 (Kotlin InputStream.use{}) |
| i18n fixes | 19 (XML comment syntax in locale files) |

## Test plan
- [x] Express API: 255/255 tests passing
- [x] Kotlin unit tests: 1829/1829 tests passing
- [x] Total: 2084 tests, zero failures

+++ app/src/main/java/com/shyden/shytalk/core/chathead/VoiceWaveView.kt
@@ -17,7 +17,7 @@ class VoiceWaveView @JvmOverloads constructor(
 ) : View(context, attrs, defStyleAttr) {
 
     private val barPaint = Paint(Paint.ANTI_ALIAS_FLAG).apply {
-        color = 0x66FFFFFF.toInt()
+        color = 0x66FFFFFF
         strokeCap = Paint.Cap.ROUND
     }
